refactor: further remove duplicates
