+++ src/main/java/edu/kit/provideq/toolbox/vrp/clusterer/KmeansClusterer.java
@@ -13,8 +13,6 @@ import edu.kit.provideq.toolbox.process.MultiFileProcessResultReader;
 import edu.kit.provideq.toolbox.Solution;
 import edu.kit.provideq.toolbox.SubRoutinePool;
 import edu.kit.provideq.toolbox.meta.Problem;
-import edu.kit.provideq.toolbox.meta.ProblemType;
-import edu.kit.provideq.toolbox.meta.SubRoutineDefinition;
 import edu.kit.provideq.toolbox.meta.setting.IntegerSetting;
 import edu.kit.provideq.toolbox.meta.setting.MetaSolverSetting;
 
@@ -37,21 +35,8 @@ public class KmeansClusterer extends VrpClusterer {
     }
 
     @Override
-    public List<SubRoutineDefinition> getSubRoutines() {
-        return List.of(
-            new SubRoutineDefinition(ProblemType.VRP,
-                "How should the clusters be solved?")
-        );
-    }
-
-    @Override
     public List<MetaSolverSetting> getSettings() {
         return List.of(new IntegerSetting(CLUSTER_SETTING_NAME, "Number of Kmeans Cluster (default: 3)", 3));
-    }
-
-    @Override
-    public boolean canSolve(Problem<String> problem) {
-        return problem.type() == ProblemType.CLUSTERABLE_VRP;
     }
 
     @Override
